+++ src/nlua_hook.h
@@ -14,8 +14,8 @@ int nlua_loadHook( nlua_env env );
 
 
 /* Misc. */
-int hookL_getarg( unsigned int hook );
-void hookL_unsetarg( unsigned int hook );
+int hookL_getarg( unsigned long hook );
+void hookL_unsetarg( unsigned long hook );
 
 
 #endif /* NLUA_HOOK */
